#955: Separate release ver markers
